requirements.txt
asgiref==3.7.2
channels==4.0.0
daphne==4.0.0
django-cors-headers==4.2.0
Django==4.2.5
djangorestframework-api-key==2.3.0
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
